Split Files and Optimize
split files into partial class and optimize some functions
